[Nova] M55: Core sim/ type changes — VTimeInSec uint64 ps, Freq uint64 Hz
- sim/event.go: VTimeInSec float64 -> uint64 (picoseconds)
- sim/freq.go: Complete rewrite with integer arithmetic, Freq uint64 (Hz)
- sim/serialengine.go: Format strings %.10f -> %d
- sim/parallelengine.go: MaxFloat64 -> MaxUint64, format strings
- sim/ticker.go: Replace sentinel -1 with hasScheduledTick bool flag
- sim/eventlogger.go: Format strings
- sim/portmsglogger.go: Format strings
- modeling/eventdriven.go: pendingWakeup sentinel -1 -> math.MaxUint64
- modeling/eventdriven_builder.go: pendingWakeup sentinel -1 -> math.MaxUint64
- tracing/averagetimetracer.go: Integer arithmetic for running average
- tracing/busytimetracer.go: VTimeInSec(0.0) -> VTimeInSec(0)
- daisen/trace.go: DB float64 -> uint64 conversions in scan methods
- monitoring/monitor.go: Format string
- examples/01_print_event: Format string
- examples/02_cell_split: Picosecond time values, format strings
- examples/ping/processor.go: 2-second delay in picoseconds
- examples/tickingping: uint64 StartTimes, format strings
- mem/acceptancetests: Format strings
- noc/acceptance/test.go: Picosecond-to-second conversion for bandwidth

+++ v5/examples/tickingping/state.go
@@ -14,7 +14,7 @@ type pingTransactionState struct {
 
 // State contains mutable runtime data for the tickingping component.
 type State struct {
-	StartTimes          []float64              `json:"start_times"`
+	StartTimes          []uint64               `json:"start_times"`
 	NextSeqID           int                    `json:"next_seq_id"`
 	NumPingNeedToSend   int                    `json:"num_ping_need_to_send"`
 	PingDst             sim.RemotePort         `json:"ping_dst"`
